voice/Quart/requirements.txt
quart
openai
azure-cognitiveservices-speech
python-dotenv